Refactor spec
