+++ spec/spec_helper.rb
@@ -28,6 +28,7 @@ require 'mi'
 
 RSpec.configure do |config|
   config.include UtilsInclude
+  config.extend UtilsExtend
   # rspec-expectations config goes here. You can use an alternate
   # assertion/expectation library such as wrong or the stdlib/minitest
   # assertions if you prefer.
