+++ spec/support/utils.rb
@@ -8,3 +8,31 @@ module UtilsInclude
     is_asserted_by{ got.include? expected }
   end
 end
+
+module UtilsExtend
+  shared_examples 'should_valid_as_a_ruby_script' do
+    it 'should valid as a ruby script' do
+      subject
+      rb = File.read(last_migration_file)
+      rip = Class.new(Ripper) do
+        def on_parse_error(msg)
+          raise msg
+        end
+      end
+      rip.parse(rb)
+    end
+  end
+
+  shared_context 'With doing', :with_doing do
+    let(:test_case){Class.new(Rails::Generators::TestCase)}
+    subject{test_case.new(:doing).run_generator(arguments)}
+    before do
+      test_case.tests described_class
+      test_case.destination_root = SPEC_TMP_DPR
+    end
+
+    after do
+      FileUtils.remove_entry_secure(Pathname.new(SPEC_TMP_DPR)/'db/migrate/')
+    end
+  end
+end
